lsusb: fix undefined behavior in convert_le_u32() signed shift
buf[3] is promoted from unsigned char to int (signed) before the left
shift. When buf[3] >= 128, shifting left by 24 produces a value that
exceeds INT_MAX, which is undefined behavior per C99/C11 6.5.7p4.

While this works on all mainstream compilers, sanitizers flag it and
optimizing compilers are permitted to exploit UB in surprising ways.

Cast buf[3] to unsigned int before shifting to ensure defined behavior.

+++ lsusb.c
@@ -154,7 +154,7 @@ static void dump_billboard_device_capability_desc(libusb_device_handle *dev, uns
 
 static unsigned int convert_le_u32 (const unsigned char *buf)
 {
-	return buf[0] | (buf[1] << 8) | (buf[2] << 16) | (buf[3] << 24);
+	return buf[0] | (buf[1] << 8) | (buf[2] << 16) | ((unsigned int)buf[3] << 24);
 }
 
 static unsigned int convert_le_u16 (const unsigned char *buf)
